continue to integrate. Creating lobby and play should be ready, but still untested
